requirements.txt
streamlit
pandas
plotly.express
scikit-learn